Add command to install static URL from ONIE
Ticket: 0
Reviewed By: no one
Testing Done:

From the ONIE shell add a command "install_url" to install a static
URL.

This command first disables the discovery process (which is
probably looping without finding anything) and then runs the ONIE
installer with the specified URL.

This is a debug/development command.  It could also come in handy from
a rescue boot.