Update u-boot ONIE version during ONIE udpate
Ticket: 0
Reviewed By: no one
Testing Done:

When updating ONIE the u-boot environment is preserved.

It's is necessary to update the onie_version environment variable
during ONIE update in order for the new version to be visable to
fw_printenv.

The new installed u-boot *does* contain the new version string in its
compiled in env, but since we're not blowing the env away as part of
the udpate we have to set it separately.