Generate LSB_RELEASE tag based on MACHINEROOT
Ticket: 0 
Reviewed By: dustin
Testing Done:

Base the LSB_RELEASE git hash on the MACHINEROOT, which may have been
specified on the command line.  This helps projects where ONIE is
housed as git submodule.  The release tag would be based on the parent
git repo (where the MACHINEROOT directory is), rather than on the ONIE
core git HEAD.

If ONIE is a submodule of the parent repo then the parent repo would
have a commit identifying the exact ONIE release, so no information is
lost.

In the case of a tarball release or other SCM scenario the user can
specify the GIT_HEAD variable on the 'make' command line.