Short circuit TFTP waterfall on a timeout
Ticket: 0
Reviewed By: scotte, dustin

If the server IP is wrong the busybox tftp client will timeout,
printing a message to stderr.

When doing a TFTP waterfall there is no point in going through the
entire waterfall on a timeout -- all the waterfall attempts will
timeout.

Note: The server IP being correct and the filename being wrong is not
a problem.  In that case the tftp client returns right away with "file
not found" -- that is how the waterfall works.

The code was tested with good and bad TFTP server IPs.  Also both
cases were excercised with and without verbosity enabled.