Change default installer file name
Reviewed By: matt, dustin

The default installer file name is changed from a single file to a
list of files, with descreasing precision.  Formerly the default
installer name was:

  onie-installer-<platform>-<arch>

In most place the single file is now replaced with the following list
of names:

  onie-installer-<arch>-<platform>
  onie-installer-<platform>
  onie-installer-<arch>
  onie-installer

This allows for simple installs with minimal infrastructure
configuration and without exact knowledge of the platform name or
architecture.

The list of file names is used for locally attached storage (USB) and
for the inexact HTTP URL methods.

The TFTP PXE-like waterfall uses onie-installer-<arch>-<platform>
during the cascade, but at the end uses the 4 default names at the
root of the server.

See the updated documentation for details and examples.