Add support for hardware revisions
Reviewed By: dustin

Add support to handle revisions of *nearly* identical hardware.

When compiling a particular $(MACHINE) the $(VENDOR_REV) variable must
also be set.  The VENDOR_REV is mapped to an internal "ONIE revision"
number, which is included in the file names of all the build products.

The idea is that several VENDOR_REVs may map to a single ONIE revision.

The fundamental ONIE platform name becomes:

  <arch>-<machine>-<machine_revision>