onie - add support for UBIFS and JFFS2
Ticket:
Reviewed By: curt
Testing Done: created some ubifs filesystems

Add mkfs.ubifs and mkfs.jffs2 to onie sysroot.

Compile time dependencies are zlib, lzo, and libuuid from e2fsprogs.  I also
had to enable the "SUSV4 legacy" option to uclibc for utime.h (for LZO)

Runtime dependencies are libz.so

This increases the sysroot.cpio.xz size by 100k bytes

Compiling binaries using the ELDK and uClibc is frowned upon.  The consensus is
that it's not always possible to convince gcc to link only uclibc libraries in
all cases.  In this change, I was able to build working biaries using the
--sysroot option to gcc and additions to the link path.  The alternative was to
replace the onie toolchain, which we decided to avoid for now.