Add dropbear ssh server and utilities
Reviewed By: dustin

Add the dropbear ssh server and client utilities to the ONIE
initramfs.

At boot time the dropbear server is started.

On the first boot the server RSA and DSS keys are generated
and stored in U-Boot environment variables.  Subsequent boots
use the stored keys.

The following commands are now available from ONIE:

  dbclient
  ssh (an alias for dbclient)
  dropbearkey -- key generator
  scp

From the outside the user can ssh to the ONIE machine as root
without using a password.  This follows the current behaviour
of the telnet server, which also does not prompt for a password.