logging: print to stdout from ssh or telnet
The installer and updater logging was going to /dev/console, which
worked great if you were on the serial console.  If you were logged in
via ssh or telnet, you would not see any output from the install_url
and update_url commands.  The output was going to /var/log/onie.log
though.

This update sends the installer/updater output to stdout.

Also in the special cases of running install_url and update_url the
output of wget and tftp are also send to stdout.  Useful when using
install_url/update_url interactively and you typed the URL wrong.