Send installer output to stdout or /dev/console
Launching an OS installer can be started in one of two ways:

1. Automatically by the image discovery process
2. User driver by the install_url CLI command

In case 1 the installer output was only going to the /var/log/onie.log
file.  It was not printed to the console.

In case 2 the installer output was correctly going to stdout and
/var/log/onie.log.  This works whether the CLI command is issued
directly from the serial console, or from an ssh or telnet session.

This patch sends the installer output to /dev/console for case 1.
This can be useful or comforting if a user is watching the automated
install process from the serial console.

Testing Done:
Exercised the various cases using the fsl_p2020rdbpca reference
platform.