u-boot: remove #${platform} from bootm command
The previous u-boot command looked like:

  bootm ${loadaddr}#${platform}

That syntax is for FIT images.  However, ONIE has not been using FIT
images until recently.  With the legacy image format the #${platform}
was just silently ignored.

With the recent use of FIT images, however, the #${platform} does have
an effect.  It puts an unnecessary constraint on the configuration
names in the FIT image.  In fact leaving it off entirely will use the
default configuration within the FIT image, which is exactly what we
want.

Also modify the installer to update the onie_bootcmd environment
variable during ONIE updates.  By design ONIE updates do not blow away
the environment as it contains variables particular to the installed
NOS.  After the ONIE update the NOS must still be able to boot.

In this case it is important to update the onie_bootcmd variable to be
in sync with the update image.