Added onie-server
This program is a DHCP server, web server and HTTP proxy all rolled into
one.  The goal of this program is help boot the OS on a switch that uses
ONIE to install its operating system.  Once the user has installed
python on their laptop, desktop, etc they can run this program which
launches the servers needed by ONIE to install an OS on a bare-metal
switch.