U-Boot pass environment size to Linux via .dtb
Ticket: ONIE-15
Reviewed By: shm

Fix up the device tree before booting by storing CONFIG_ENV_SIZE in
the device tree.

First find the FDT node with a property called "label" whose value is 
"uboot-env". This corresponds to the NOR flash partition used for the 
U-Boot environment variables. 

If the node is found then set the "env_size" property to 
CONFIG_ENV_SIZE. 

On the other side the NOS reads the value from the device so it knows 
how to configure the fw_printenv and fw_setenv tools. 

This is currently a problem with Accton's U-Boot.

Testing Done:

Compiled a kernel/dtb/initramfs with env_size set to a random number
in the dtb.  I then boot the image from U-Boot and verified that
env_size was set correctly in the .dtb and that fw_setenv and
fw_printenv continued to work.