build system: Add local.make mechanism
Added a mechanism where the top level Makefile looks for "local.make"
and includes it if it exists.

Setting variables in "local.make" file allows the user to override
various defaults set in the other makefiles.  The other makefiles set
some variables using the "=?" operator, only setting the variable if
it is not already defined.

The local.make.example contains a number of example over rides.