Allow DHCP to configure interfaces every iteration
The original concept of locking out DHCP interface configuration once
a fall back IP address is configured was flawed.

The problem happens when the DHCP server is unresponsive during boot
up and ONIE uses a default IP configuration.  Later during the
discovery loop if the DHCP server comes back up ONIE will continue to
use the default IP config, i.e. it will not update the Ethernet
management interfaces again.

The only way to recover is for ONIE to reboot and go through the boot
up sequence again.

The fix is to remove the "fallback active" tracking.  The fallback IP
address is only used when the DHCP server is unresponsive.

Testing:

1. Brought up the DUT and let DHCP configure the interface

2. Disabled the DHCP server

3. The next discovery loop ONIE fell through to using the fallback IP
   address

4. Re-enable the DHCP server

5. The next discovery loop ONIE uses the IP address from the DHCP
   server

Reviewed By: Vidya Ravipati <vidya@cumulusnetworks.com>
Signed-off-by: Curt Brune <curt@cumulusnetworks.com>