Mount ONIE tools from running NOS
Problem: An installed NOS needs to be able to use the onie-grub-XXX
CLIs.

This patchs adds a new script, mount_onie.sh, to the ONIE image that
lives in the ONIE-BOOT partition.  This script mounts the ONIE-BOOT
initramfs on a tmpfs that the NOS can access.

This patch also adds an example CLI to the demo NOS, /bin/mount-onie,
showing how to setup and call the ONIE-BOOT/mount_onie.sh script.