Use fakeroot instead of sudo
On the build host, instead of installing the packages using 'sudo',
create the final initrd using a 'fakeroot' environment.  This avoids
the use of sudo and running things as root on the build host.

This patch updates the various package Makefiles to remove the use of
'sudo'.  The patch also adds helper script, make-sysroot.sh, which
creates the final initrd using a fakeroot environment.