onie-boot-default: ONIE mode CLI option
The aim is to make specifying the various ONIE modes easier.

Previously to use this command you need to know the *extact* menu
entry text used in the grub configuration file.  This string can be
arbitrarily complex, including all the paths through a grub submenu.

This patch adds a new option "-o" that lets the user specify one of
the supported "ONIE modes".  The available ONIE modes are:

  install     -- ONIE OS installer mode
  rescue      -- ONIE rescue mode
  uninstall   -- ONIE OS uninstall mode
  update      -- ONIE self update mode

Internally the simple string is converted to the actual GRUB menu
entry, which could be long and complex.

Reviewed-By: dustin