discovery: optionally skip some local block device partitions
Add a general mechanism for skipping local block device partitions
during the discover process.  An architecture can define
rootconf/$(ARCH)/sysroot-lib-onie/discover-arch, which contains a list
of partitions to skip:

In that file an architecture defines the skip_parts_arch() function,
which echos a list of partition identifiers to skip.  The core
discovery mechanism will skip any partitions where the output of the
`blkdid' command matches these strings.

The genesis for this need is the x86_64 architecture where ONIE uses
some of the local partitions for itself.  It makes no sense to mount
the ONIE-BOOT and ONIE-CONFIG partitions for this architecture, as an
installer image will never be located there.

Reviewed-By: dustin