onie-tools: Create a tarball of tools, for use by a NOS
An installed NOS needs a way to communicate with ONIE via the
ONIE-CONFIG partition.  This patch adds a set of tools as a tarball to
the read-only ONIE-BOOT partition.

The idea is a NOS would mount the ONIE-BOOT partition read-only and
install the tools for use.

The tools provided are:
	onie-boot-default
	onie-boot-entry-add
	onie-boot-entry-remove
	onie-boot-entry-show
	onie-boot-update

The tools are kept in a xz compressed tar ball.

The tools need to be usable by a wide variety of operating systems.
To this end .tar.xz format was used thinking it would be easy for OS
to install the tools.

A helper script, mount-onie-tools, is provided to easily install the
tools in a tmpfs for the NOS.

Reviewed-By: dustin