ONIE installer: allow platforms to define the ONIE install block device
The problem we are trying to solve is:

   How to determine the block device upon which to install ONIE?

The question is complicated when multiple block devices are present,
i.e. perhaps the system has two hard drives installed or maybe a USB
memory stick is momentarily installed.  For example the internal disk
device usually shows up as /dev/sda under Linux, but maybe with a USB
drive connected the internal disk now shows as /dev/sdb.

This patch forces each platform to specify how to determine the target
ONIE install device.