ONIE uninstaller: updates to make uninstall architecture independent
This patch forces each architecture to specify how to "clean up" the
machine.

An architecture must provide an implementation for the
uninstall_system() functions.  Optionally a platform may override the
architecture implementation.

The core uninstaller expects to find the files here:
  [ -r $lib_dir/uninstall-arch ]     && . $lib_dir/uninstall-arch
  [ -r $lib_dir/uninstall-platform ] && . $lib_dir/uninstall-platform

Finally this patch also renames the uninstaller CLI to onie-uninstaller.
A symlink is made from onie-uninstaller to the old name to preserve
legacy systems.