EEPROM-TLV-Add-new-service-tag-TLV-code
Reviewed By: Curt Brune <curt@cumulusnetworks.com>

This patch adds a new "Service Tag" TLV to the ONIE EEPROM format.

The new TLV code is 0x2F.

The format of the data is a variable length ASCII string.

It is intended to contain a vendor defined service tag.

Testing Done:

Updated an existing system and exercised the new TLV:

- verified setting the new TLV:
    sys_eeprom set 0x2F "fake service tag"

- verified displaying the new TLV:
    sys_eeprom

- verified removing the new TLV:
    sys_eeprom set 0x2F

- verified the new code shows up in the list of available codes:
    sys_eeprom list

Signed-off-by: Srideep <srideep_devireddy@dell.com)>