onie-sysinfo: allow platform and architecture overrides
Reviewed By: vidya, dustin

The onie-sysinfo command dumps a number of different information
fields.

This patch lets a particular implementation override the default
beahviour for the 'serial number' and 'base MAC address' fields.

For both of these fields search for the value in the following order:

1. is the value set in the environment?  highest priority
2. the platform provides an accessor function
3. the CPU architecture provides an accessor function
4. use the contents of /sys/class/net/eth0/address
5. return 'unknown'

Setting the value in the environment has the highest priority.  This
is useful in development where you can set the value on the kernel
command line.

Next, a platform implementation takes priority.  If a particular
platform has special hoops to jump through it can implement a
function.

Next comes the CPU architecture implementation.  This implementation
should apply to a wide class of machines that share a common
access method.

Testing Done:

Exercised the implementation on PowerPC and x86_64 platforms.
Exercised the priority order, testing each one in turn.