x86_64: onie-sysinfo arch overrides
Reviewed By: vidya, dustin

Provide x86_64 arch overrides for the onie-sysinfo 'serial number' and
'base MAC address' fields.

On x86_64 the default implementation will rely on the onie-syseeprom
command and the ONIE TLV EEPROM format to locate the required
information.

Testing Done:

Exercised 'onie-sysinfo -a' on a x86_64 based system and validated the
results.