gptfdisk: use exported ONIE compiler and linker flags
Reviewed By: puneet

Updates the gptfdisk Makefile to append the ONIE compiler and linker
flags to its own definitions.

Previously the ONIE gptfdisk.make commands were overriding the
gptfdisk project's definitions with the ONIE definitions.  We really
wanted the ONIE definitions to be appended to the project's
definitions, not to replace them.

Testing Done:
Compiled images and exercised on x86 platform