onie-syseeprom:  add get TLV by type_code option
Reviewed By:  puneet, scotte, dustin

This patch adds a new option, -g (--get) <TLV code>, that extracts the
relevant TLV from the EEPROM and prints it to stdout.  This option is
very nice for scripts that are interested in particular fields, like
finding the serial number:

  ONIE:/ # /bin/onie-syseeprom -g 0x23
  fake-serial-12345

Using this option scripts can extract specific fields without dealing
with error prone screen scraping techniques.

This patch also cleans up a number of error handling cases that
previously went undetected.  The exit status of this program now
accurately reflects the status of the operation.

Testing Done:

- ran with -h, verifying the new option is described correctly
- ran with -g for existing code. verified dumped code matched what was
  in the eeprom.  Program return code was 0.

- ran with -g for non-existing code.  verified error reported and
  program return code was non-zero.

- exercised existing features, like -s, -l, -e, verified the program
  return status was 0.

- exercised existing -e with bad code and verified the program return
  status was non-zero.