kvm_x86_64:  create virtio device dynamically
Commit 062d8d29e88e5fed197fbe5628311016ba522567 disabled the CONFIG_MD
kernel parameter for this platform because the virtio devices were not
working properly.

The cause of the problem is that the virto MAJOR device number is
determined dynamically, while we were assigning it statically.

This patch reverts the CONFIG_MD kernel configuration change and
creates the virtio devices, /dev/vda[1-8], dynamically at boot time.

Testing:

- Exercised the image on the kvm_x86_64 target, which uses virtio.

- Exercised the image on a real x86_64 target, which does not use
  virtio.