Check if network interface is operationally "up" before configuration
This patch adds a check to the Ethernet management interface(s) for
"link up" before trying to configure the interface via DHCP.

On platforms with multiple management interfaces this check is a boot
time optimization.  In those cases where only one of the interfaces
has link, the boot process would pause waiting for the unconnected
interface DHCP operation to time out.  With this patch the boot
process does not pause, moving on after configuring the first
interface.

The check tries to verify the link up condition for 5 seconds before
declaring the link is down.

Reviewed By:  dustin

Testing Done:

Built images and verified the boot time 'link up' detection process.

- The process was verified for PowerPC and x86 platforms.

- For each platform the process was verified with port speeds 1000M,
  100M and 10M.