demo: fix build time race condition on sysroot
The building of the demo sysroot depended on the
$(SYSROOT_COMPLETE_STAMP) stamp.  However, a few files in the sysroot
are modified as part of the sysroot CPIO.xz creation process.  In
particular the step that uses fakeroot to make the /dev files
continues to modify files.

Under certain parallel build conditions the demo sysroot target would
fail as the main sysroot files were being modified underneath.

This patch makes the demo sysroot depend on the final sysroot CPIO.xz,
eliminating the race condition.