popt: clean up build dependencies
The POPT_NEW_FILES variable captures all the files in the popt
directory that are more recent than the $(POPT_BUILD_STAMP) stamp.

The check should be made for the popt-build target, however it was
mistakenly being made for the popt-config target.  As a result during
re-compiles popt and gptfdisk (upon which popt depends) would
unnecessarily be rebuilt.

This patch corrects the new file detection.