grub2: update to newer version
In order to support PXE-EFI64 booting we need a more up to date
version of GRUB2.

This patch moves the GRUB version to 2.02~beta2+e4a1fe391.

The e4a1fe391 refers to the git hash from the master grub repo:
  git clone git://git.savannah.gnu.org/grub.git

The 2.02~beta2 release did not work so we moved to the latest
available.

Two minor patches are added on top:

version-string.patch:
change the version string to "2.02~beta2+e4a1fe391"

set-default-graphics-mode-to-text.patch:
For EFI64 set the default graphics mode to text