grub-host: compile grub for build host
In order to build PXE EFI64 images we need to run grub locally on the
build host.

This patch adds a new build target, grub-host, which is built when
building platforms that enable PXE EFI64 images.

The grub-host target is configured and compiled using the build host's
toolchain and libraries.

This is in contrast to the grub target, which is configured and
compiled using the target platform's toolchain.