Create /usr/share/locale in the sysroot
Trivially create /usr/share/locale directory in the ONIE sysroot.

The updated GRUB2 binaries are expecting to find it and print a
harmless, yet annoying, error message when it is missing.

Previously only the recovery sysroot was creating this directory.
This patch moves the directory creation into the main sysroot, which
used by both the regular ONIE updater image and the recovery image.