onie-syseeprom: use correct endianness for the data length
Reviewed By: Scott Emery <scotte@cumulusnetworks.com>

The ONIE TLV EEPROM format specifies a TLV header, which contains a
two byte length field describing the overall length of the EEPROM
contents.  According to spec this value *must* be stored in big endian
format.  See the 'Definition of the TlvInfo EEPROM Format' section in
onie/docs/design-spec/hw_requirements.rst for more on that.

The original code was mistakenly including
<linux/byteorder/big_endian.h> instead of the more general
<asm/byteorder.h> for the __be16_to_cpu(x) and __cpu_to_be16(x)
macros.  Using big_endian.h directly makes these macro no-ops as in
this case the CPU is assumed to be big endian.

By using the more general <asm/byteorder.h> from the toolchain's
sysroot the correct endiannes of the target CPU is used.

Testing Done:

Hexdumps of the EEPROM data before and after this change confirmed the
byte swapped length field.  Before the fix the mis-encoded EEPROM data
was not decodable by other programs which implemented the TLV standard
correctly.  After the fix, those programs worked fine.