onie-syseeprom: print warnings for bad EEPROM contents
Reviewed By: Scott Emery <scotte@cumulusnetworks.com>

When onie-syseeprom detects invalid EEPROM contents it starts off with
a in-ram copy of default contents.  The default contents consists of a
the TLV header info and the CRC TLV.

What is missing is an indication to the user *why* the default
contents are being used.

This patch adds warning messages when the TLV header is corrupt or
when the EEPROM checksum is bad.

Testing Done:

With a blank EEPROM the onie-syseeprom command responds as follows:

  ONIE:/ # onie-syseeprom
  Notice:  Invalid TLV header found.  Using default contents.
  Notice:  Invalid TLV checksum found.  Using default contents.
  TlvInfo Header:
     Id String:    TlvInfo
     Version:      1
     Total Length: 6
  TLV Name             Code Len Value
  -------------------- ---- --- -----
  CRC-32               0xFE   4 0xD4431C18