Use virtual machine to mimic a real platform
Reviewed By: tbd

This patch introduces two optional Makefile variables that can be used
to override the MACHINE and PLATFORM variables when creating the
runtime /etc/machine.conf file.

The new variables are:

  RUNTIME_ONIE_PLATFORM
  RUNTIME_ONIE_MACHINE

This is particularly useful when MACHINE is the kvm_x86_64 virtual
machine.  Using these variables you can make the running virtual
machine look like a specific real machine.  This is useful when
developing an installer for a particular platform.  You can develope
the installer using the virtual machine.

For example to compile the virtual machine to look like the Interface
Master im_n29xx_t40n machine you would build the image like this:

  build:~/onie/build-config$ make -j8 MACHINE=kvm_x86_64 RUNTIME_ONIE_MACHINE=im_n29xx_t40n all

The net effect is that the onie-sysinfo command will report that the
machine is im_n29xx_t40n while running in a virtual machine.

Since real NOS installers use onie-sysinfo to determine the running
machine this mechanism can be used to perform initial OS installer
development for a new platform.