image discovery x86:  make sure onie_eth_addr is set
Reviewed By: Nike Chen <nike_chen@accton.com>, David Yang <david_yang@accton.com>

During the execution of /bin/exec_installer the variable
'onie_eth_addr' is used in a few places.  On PowerPC platforms this
variable is set on the kernel command line when U-Boot loads and boots
ONIE.

On x86 platforms this variable was not being set correctly.  This
patch corrects that.

If the onie_eth_addr variable is unset then set it using the output of
"onie-sysinfo -e", which returns the MAC address for eth0.

Testing Done:

Compiled images for x86 and PowerPC.  Verified the TFTP waterfall contains
a URL of form:

  tftp://<server>/<MAC_Address>/onie-installer-x86_64-<vendor>_<machine>-r<machine_revision>

Also verified that HTTP requests are setting the ONIE-ETH-ADDR: HTTP
header correctly.