Remove 32-bit package from debian-prepare-build-host target
The debian-prepare-build-host Makefile target installs needed packages
to prepare a build host for building ONIE.

Along the way we picked up a 32-bit package, which does not work when
the build host is x86_64.

The intent of the build target is more for guidance and reference than
to provide an absolute rule.  For now make the assumption the build
host is a x86_64 machine.