Add optional vendor version string to final ONIE version string
This patch enables each machine to define an optional "version string"
that is appended to the final ONIE version string.

This allows hardware vendors to append a machine specific string to
the end of the version string.

An example of how to use this feature is in the kvm_x86_64 virtual
machine's machine.make file:

  machine/kvm_x86_64/machine.make

The contents of the optional $(VENDOR_VERSION) variable are appended
to the final ONIE version string.

See the comments in build-config/Makefile for a discussion of how the
ONIE_RELEASE_TAG (internal ONIE version string) and LSB_RELEASE_TAG
(external, final version string) are constructed.