Move uClibc version selection into crosstool-ng
For historical reasons the ONIE project downloaded uClibc-0.9.32.1
outside of the crosstool-ng build process.  The crosstool-ng
configuration listed the uClibc version as "custom" and the custom
location of the tarball was specified separately.

That was fine for a start.

In preparation for opening up ONIE to other versions of uClibc the old
way does not scale anymore.

This patch does a few things:

1.  Specify the desired uClibc version *within* the crosstool-ng
    configuration file.  Allows crosstool-ng to manage the uClibc
    version.

2.  Move the existing crosstool-ng CPU architecture configuration
    files to a new location, based on the uClibc version.