demo os: fix mounting of demo partition
Reviewed By: Will_Kuo@alphanetworks.com

Commit 6d7cc40213cba8d5a5e63f6958a0220b92d1f791 broke the ability of
the demo OS to mount the "DEMO" partition.  This was broken for all
x86 platforms.

This patch restores the mounting of the ONIE demo OS partition on
/boot at demo boot time.

Instead of relying on the old rc.local mechanism, move the mount
script to /etc/rcS.d/S05mount-demo.sh for x86 platforms.

Testing Done:

Built ONIE and DEMO-OS images for the kvm_x86_64 machine and verified
that when the demo OS boots the ONIE-DEMO-OS partition is mounted:

  KVM_X86_64-OS:/etc/rcS.d # blkid
  /dev/vda3: LABEL="ONIE-DEMO-OS" UUID="c05da06a-151b-4659-bd1b-6785b3f23248"
  /dev/vda2: LABEL="ONIE-BOOT" UUID="db340634-168d-4d86-ba8a-9236d0f815c3"
  KVM_X86_64-OS:/etc/rcS.d # df
  Filesystem           1K-blocks      Used Available Use% Mounted on
  tmpfs                    10240        12     10228   0% /run
  tmpfs                    10240         0     10240   0% /run/lock
  tmpfs                   253288         0    253288   0% /tmp
  tmpfs                   253288         0    253288   0% /var/tmp
  tmpfs                   253288         0    253288   0% /run/shm
  /dev/vda2               122835     12107    101554  11% /mnt/onie-boot
  /dev/vda3               122835     11813    101848  10% /boot